Issues-6: Moved RecyclerSearchActivity to separated module.
