+++ app/build.gradle
@@ -62,4 +62,5 @@ dependencies {
     implementation project(path: ':recyclermvp')
     implementation project(path: ':coordinatorbehavior')
     implementation project(path: ':countdownlabel')
+    implementation project(path: ':recyclersearch')
 }
