+++ recyclersearch/src/main/res/layout/a_recycler_search.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ui.recycler.search.RecyclerSearchActivity">
+	tools:context="com.kirchhoff.recyclersearch.RecyclerSearchActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
