+++ recyclersearch/src/main/res/layout/r_card.xml
