+++ recyclersearch/src/main/res/menu/recycler_search.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".ui.recycler.search.RecyclerSearchActivity">
+	tools:context="com.kirchhoff.recyclersearch.RecyclerSearchActivity">
     <item
         android:id="@+id/search"
         android:icon="@drawable/ic_search"
