+++ recyclersearch/src/main/res/values/strings.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+	<string name="recyclerview_with_search_title">RecyclerView with search</string>
+	<string name="search">Search</string>
+</resources>
